Extract shared ConnectorSegment composable to eliminate duplication
Move connector colors, labels, and ConnectorSegment composable into
WorkflowNodeOrder.kt so both WorkflowNodeCard and
WorkflowTemplateDetailScreen use the same component.
