+++ app/src/main/kotlin/io/github/leogallego/ansiblejane/ui/workflows/WorkflowNodeOrder.kt
@@ -1,9 +1,82 @@
 package io.github.leogallego.ansiblejane.ui.workflows
 
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.drawBehind
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.StrokeCap
+import androidx.compose.ui.unit.dp
 import io.github.leogallego.ansiblejane.model.WorkflowJobTemplateNode
 import io.github.leogallego.ansiblejane.model.WorkflowNode
 
 enum class ConnectorType { SUCCESS, FAILURE, ALWAYS }
+
+private val connectorSuccessColor = Color(0xFF4CAF50)
+private val connectorFailureColor = Color(0xFFF44336)
+private val connectorAlwaysColor = Color(0xFF9E9E9E)
+
+fun ConnectorType.color(): Color = when (this) {
+    ConnectorType.SUCCESS -> connectorSuccessColor
+    ConnectorType.FAILURE -> connectorFailureColor
+    ConnectorType.ALWAYS -> connectorAlwaysColor
+}
+
+fun ConnectorType.label(): String = when (this) {
+    ConnectorType.SUCCESS -> "On success"
+    ConnectorType.FAILURE -> "On failure"
+    ConnectorType.ALWAYS -> "Always"
+}
+
+@Composable
+fun ConnectorSegment(type: ConnectorType) {
+    val color = type.color()
+    val label = type.label()
+
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(horizontal = 16.dp),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Box(
+            modifier = Modifier
+                .width(32.dp)
+                .height(32.dp)
+                .drawBehind {
+                    val centerX = size.width / 2f
+                    drawLine(
+                        color = color,
+                        start = Offset(centerX, 0f),
+                        end = Offset(centerX, size.height),
+                        strokeWidth = 3.dp.toPx(),
+                        cap = StrokeCap.Round
+                    )
+                    drawCircle(
+                        color = color,
+                        radius = 4.dp.toPx(),
+                        center = Offset(centerX, size.height / 2f)
+                    )
+                }
+        )
+        Spacer(modifier = Modifier.width(4.dp))
+        Text(
+            text = label,
+            style = MaterialTheme.typography.labelSmall,
+            color = color
+        )
+    }
+}
 
 data class OrderedNode(
     val node: WorkflowNode,
