+++ app/src/main/kotlin/io/github/leogallego/ansiblejane/ui/workflows/WorkflowTemplateDetailScreen.kt
@@ -1,15 +1,12 @@
 package io.github.leogallego.ansiblejane.ui.workflows
 
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.material.icons.Icons
@@ -28,10 +25,6 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.drawBehind
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.StrokeCap
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
@@ -39,22 +32,6 @@ import io.github.leogallego.ansiblejane.presentation.workflows.WorkflowTemplateD
 import io.github.leogallego.ansiblejane.presentation.workflows.WorkflowTemplateDetailViewModel
 import io.github.leogallego.ansiblejane.ui.components.ErrorMessage
 import org.koin.compose.viewmodel.koinViewModel
-
-private val connectorSuccessColor = Color(0xFF4CAF50)
-private val connectorFailureColor = Color(0xFFF44336)
-private val connectorAlwaysColor = Color(0xFF9E9E9E)
-
-private fun ConnectorType.color(): Color = when (this) {
-    ConnectorType.SUCCESS -> connectorSuccessColor
-    ConnectorType.FAILURE -> connectorFailureColor
-    ConnectorType.ALWAYS -> connectorAlwaysColor
-}
-
-private fun ConnectorType.label(): String = when (this) {
-    ConnectorType.SUCCESS -> "On success"
-    ConnectorType.FAILURE -> "On failure"
-    ConnectorType.ALWAYS -> "Always"
-}
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -143,41 +120,7 @@ private fun TemplateNodeCard(orderedNode: OrderedTemplateNode) {
 
     Column(modifier = Modifier.fillMaxWidth()) {
         if (incomingEdge != null) {
-            val color = incomingEdge.color()
-            val label = incomingEdge.label()
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(horizontal = 16.dp),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Box(
-                    modifier = Modifier
-                        .width(32.dp)
-                        .height(32.dp)
-                        .drawBehind {
-                            val centerX = size.width / 2f
-                            drawLine(
-                                color = color,
-                                start = Offset(centerX, 0f),
-                                end = Offset(centerX, size.height),
-                                strokeWidth = 3.dp.toPx(),
-                                cap = StrokeCap.Round
-                            )
-                            drawCircle(
-                                color = color,
-                                radius = 4.dp.toPx(),
-                                center = Offset(centerX, size.height / 2f)
-                            )
-                        }
-                )
-                Spacer(modifier = Modifier.width(4.dp))
-                Text(
-                    text = label,
-                    style = MaterialTheme.typography.labelSmall,
-                    color = color
-                )
-            }
+            ConnectorSegment(type = incomingEdge)
         }
 
         ElevatedCard(
